src/main/resources/python/requirements.txt
speechrecognition==3.10.4
pydub==0.25.1
google-cloud-language==2.13.4